requirements.txt
Flask==3.0.0
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
python-docx==1.1.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
